# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for NutriApp - Nutritionist management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerNutritionist
      tags: [auth]
      summary: Register first nutritionist account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /auth/me/password:
    patch:
      operationId: changeMyPassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── Dashboard ──────────────────────────────────────────
...
  /biblioteca:
    get:
      operationId: listPublications
      tags: [biblioteca]
      summary: List publications (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, published, draft]
        - name: category
          in: query
          schema:
            type: string
            enum: [all, ebook, receita, video, artigo]
      responses:
        "200":
          description: Publications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publication"
    post:
      operationId: createPublication
      tags: [biblioteca]
      summary: Create publication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
  /biblioteca/{id}:
    get:
      operationId: getPublication
      tags: [biblioteca]
      summary: Get publication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Publication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "404":
          description: Not found
    patch:
      operationId: updatePublication
      tags: [biblioteca]
      summary: Update publication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
    delete:
      operationId: deletePublication
      tags: [biblioteca]
      summary: Delete publication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /biblioteca/{id}/toggle-status:
    patch:
      operationId: togglePublicationStatus
      tags: [biblioteca]
      summary: Toggle publication status (draft/published)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"

  # ── Patient Portal ─────────────────────────────────────
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth Schemas ──────────────────────────────────────
    AuthUser:
      type: object
      required: [id, email, role, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [nutricionista, paciente]
        patientId:
          type: ["integer", "null"]
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # ── Dashboard Schemas ─────────────────────────────────
...
    GroupSummary:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
...
    Publication:
      type: object
      required: [id, title, category, status, visibility, createdAt, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [ebook, receita, video, artigo]
        status:
          type: string
          enum: [rascunho, publicado]
        visibility:
          type: string
          enum: [geral, grupos, pacientes]
        videoUrl:
          type: ["string", "null"]
        pdfPath:
          type: ["string", "null"]
        imagePath:
          type: ["string", "null"]
        groupIds:
          type: array
          items:
            type: integer
        patientIds:
          type: array
          items:
            type: integer
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        createdAt:
          type: string
        publishedAt:
          type: ["string", "null"]
        isNew:
          type: boolean
    PublicationInput:
      type: object
      required: [title, category, visibility]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
          enum: [ebook, receita, video, artigo]
        status:
          type: string
          enum: [rascunho, publicado]
        visibility:
          type: string
          enum: [geral, grupos, pacientes]
        videoUrl:
          type: string
        pdfPath:
          type: string
        imagePath:
          type: string
        groupIds:
          type: array
          items:
            type: integer
        patientIds:
          type: array
          items:
            type: integer
    PublicationUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [ebook, receita, video, artigo]
        status:
          type: string
          enum: [rascunho, publicado]
        visibility:
          type: string
          enum: [geral, grupos, pacientes]
        videoUrl:
          type: ["string", "null"]
        pdfPath:
          type: ["string", "null"]
        imagePath:
          type: ["string", "null"]
        groupIds:
          type: array
          items:
            type: integer
        patientIds:
          type: array
          items:
            type: integer

    # ── Storage Schemas ───────────────────────────────────
...